chore(GiniBankAPILibrary): Add tests for GiniBankAPILibrary with pinning configuration.
PIA-2485
